Add files via upload
This code is applying the Linked List Data structure into C using struct and Dynamic memory allocation. The Idea here is to have a soccer team with many attributes such as first name, last name, position and value. we could possibly have 4 positions, Goalkeeper, Defender, Midfielder and striker. when a member is added to the list we must add it to the end of its similar position while reserving the order as mentioned before(G->D->M->S). many other functions are included here to do many other operations to the list.